# docs/knowledge/backend/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      summary: Request OTP for login (all roles)
      tags: &a1
        - Auth
  /api/auth/verify-otp:
    post:
      operationId: AuthController_verifyOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpDto"
      responses:
        "201":
          description: ""
      summary: Verify OTP and login (all roles)
      tags: *a1
  /api/auth/me:
    get:
      operationId: AuthController_getProfile
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Get current user profile
      tags: *a1
...
  /api/analytics/super-admin/overview:
    get:
      operationId: AnalyticsController_getSystemOverview
      parameters: []
      responses:
        "200":
          description: ""
      security: &a16
        - access-token: []
      summary: Get system-wide overview metrics for Super Admin
      tags: &a17
        - Analytics
  /api/analytics/super-admin/visitor-trends:
    get:
      operationId: AnalyticsController_getVisitorTrends
      parameters:
        - name: period
          required: false
          in: query
          schema:
            enum:
              - daily
              - weekly
              - monthly
              - yearly
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visitor trends (daily/weekly/monthly/yearly)
      tags: *a17
  /api/analytics/super-admin/visit-status-distribution:
    get:
      operationId: AnalyticsController_getVisitStatusDistribution
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visit status distribution
      tags: *a17
  /api/analytics/super-admin/visit-category-distribution:
    get:
      operationId: AnalyticsController_getVisitCategoryDistribution
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visit category distribution (Meeting vs Delivery)
      tags: *a17
  /api/analytics/super-admin/user-role-distribution:
    get:
      operationId: AnalyticsController_getUserRoleDistribution
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get user role distribution
      tags: *a17
  /api/analytics/super-admin/chains/stats:
    get:
      operationId: AnalyticsController_getAllChainsWithStats
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get all chains with their statistics
      tags: *a17
  /api/analytics/super-admin/branches/stats:
    get:
      operationId: AnalyticsController_getAllBranchesWithStats
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get all branches with their statistics
      tags: *a17
  /api/analytics/super-admin/chain/{chainId}/stats:
    get:
      operationId: AnalyticsController_getChainStats
      parameters:
        - name: chainId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get statistics for a specific chain
      tags: *a17
  /api/analytics/super-admin/branch/{branchId}/stats:
    get:
      operationId: AnalyticsController_getBranchStats
      parameters:
        - name: branchId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get statistics for a specific branch
      tags: *a17
  /api/analytics/super-admin/chains/growth:
    get:
      operationId: AnalyticsController_getChainGrowth
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get chain growth over time
      tags: *a17
  /api/analytics/super-admin/branches/growth:
    get:
      operationId: AnalyticsController_getBranchGrowth
      parameters: []
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get branch growth over time
      tags: *a17
  /api/analytics/chain-admin/overview:
    get:
      operationId: AnalyticsController_getChainAdminOverview
      parameters:
        - name: chainId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get chain overview metrics for Chain Admin
      tags: *a17
  /api/analytics/chain-admin/visitor-trends:
    get:
      operationId: AnalyticsController_getChainVisitorTrends
      parameters:
        - name: chainId
          required: true
          in: query
          schema:
            type: string
        - name: period
          required: false
          in: query
          schema:
            enum:
              - daily
              - weekly
              - monthly
              - yearly
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visitor trends for a specific chain
      tags: *a17
  /api/analytics/chain-admin/branches/stats:
    get:
      operationId: AnalyticsController_getChainBranchesStats
      parameters:
        - name: chainId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get all branches with stats for a specific chain
      tags: *a17
  /api/analytics/branch-admin/visitor-trends:
    get:
      operationId: AnalyticsController_getBranchVisitorTrends
      parameters:
        - name: branchId
          required: true
          in: query
          schema:
            type: string
        - name: period
          required: false
          in: query
          schema:
            enum:
              - hourly
              - daily
              - weekly
              - monthly
              - yearly
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visitor trends for a specific branch
      tags: *a17
  /api/analytics/security/visitor-trends:
    get:
      operationId: AnalyticsController_getSecurityVisitorTrends
      parameters:
        - name: branchId
          required: true
          in: query
          schema:
            type: string
        - name: period
          required: false
          in: query
          schema:
            enum:
              - hourly
              - daily
              - weekly
              - monthly
              - yearly
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visitor trends for security personnel
      tags: *a17
  /api/analytics/staff/visitor-trends:
    get:
      operationId: AnalyticsController_getStaffVisitorTrends
      parameters:
        - name: branchId
          required: true
          in: query
          schema:
            type: string
        - name: period
          required: false
          in: query
          schema:
            enum:
              - hourly
              - daily
              - weekly
              - monthly
              - yearly
            type: string
      responses:
        "200":
          description: ""
      security: *a16
      summary: Get visitor trends for staff personnel
      tags: *a17
info:
  title: Hospital Visitor Tracking System
  description: API documentation for the Hospital Visitor Tracking System
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties:
        phone:
          type: string
          example: "6987456321"
      required:
        - phone
    VerifyOtpDto:
      type: object
      properties:
        phone:
          type: string
          example: "6987456321"
        otp:
          type: string
          example: "123456"
      required:
        - phone
        - otp
...